Updates Database for Remote Server
Updated the local database and exported sql file to match the remote server with latin1_swedish_ci collation, and updated the table and user names for accessing the database to match the requirements for the remote server. Changed some details in the auto detection of remote or local hosting in the server file.